# Conflicts:
#	src/main/java/ru/skypro/homework/service/mapper/AdMapper.java

+++ src/main/java/ru/skypro/homework/repository/AdRepository.java
@@ -4,6 +4,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.skypro.homework.entity.Ad;
 
+/**
+ * Интерфейс, содержащий методы для работы с базой данных объявлений
+ * @see ru.skypro.homework.entity.Ad
+ * @see ru.skypro.homework.service.AdService
+ * @see ru.skypro.homework.service.impl.AdServiceImpl
+ */
 @Repository
-public interface AdRepository extends JpaRepository<Ad, Long> {
+public interface AdRepository extends JpaRepository<Ad, Integer> {
 }
